sdk: fix buggy shell behaviour on unbutu after sourcing environment-setup
This fixes an issue for end users using the toolchain on ubuntu, when they
type an invalid command in the shell.

The environment setup script generated with the cross compilation toolchain
redefines the PYTHONHOME environment variable to use the local python version.

On ubuntu, typing an unknown command in a shell will run the
ubuntu-specific command_not_found_handle python script. The problem is that
this script is not compatible with the python version shipped with the cross
compilation toolchain.

Tracked-On: EDISON-2377
Signed-off-by: Fabien Chereau <fabien.chereau@intel.com>