alsa-utils: ensure that asound.state file is written on drive
When issuing the command "reboot", the systemd service alsa-store.service is
started. This service writes on drive the alsa configuration. Some times, the
file is not well written, leading to error at boot when alsa-restore service
tries to read this corrupted file.

This patch tries to solve this bug by adding a "sync" in the service.

Tracked-On: EDISON-2450
Signed-off-by: Simon Desfarges <simon.desfarges@intel.com>