FTL: enable notifier widget
Currently, when a user clicks on start to flash from FTL tool, nothing
is displayed to explain how and when to plug Edison board.

This patch enables the notifier widget for FTL tool :
when a user will click on start to flash, a notifier will be displayed
on bottom right corner, explaining how and when to connect Edison board.
It also gives some useful information about flashing status.

Tracked-On: EDISON-2488
Signed-off-by: Rodriguez Fabien <fabienx.rodriguez@intel.com>