edison.conf: change machine to core-i7 and build kernel using x86_64_defconfig
For now remove U-Boot from image as it fails to link and stalls building the rootfs as that depended on it.
It is debateable if U-Boot should be on the rootfs image itself. Moreover building i386 U-Boot with x86_64
tools leads to linking problems that probably need to be solved by adding multilib. Maybe it will be better
to build U-Boot as a seperate image by i386 tools.
Without U-Boot the kernel, initramsfs and rootfs build and boot just fine.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>
On branch morty-64

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>