u-boot: switch to 2018.01 and select acpi by DISTRO_FEATURES
With this patch for the non-acpi enabled version we switch to upstream and apply a selection of patches
for Edison. For the acpi version we use Andy Shevchenko's acpi branch. The selection is made by by setting
acpi in the DISTRO_FEATURES variable in poky-edison.conf
Note: booting non-acpi enabled kernels requires adding acpi=off on the command line

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>