setup: make life easier by adding a make update
make setup deletes the out directory and destroys the conf files each time it is run. Instead we now have a
make update command and setup creates links to the conf files in the meta-intel-edison directory. This alows
the conf files to be checked in. it also allows experimental conf files to be easily changed without changing the setup

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>