edison-image: correct live image generation config
Openembedded-core commit

5493bb5ba4b4520f944d38b214a3d53a5282e0be
live-vm-common.bbclass: Don't use vmlinuz or VM_DEFAULT_KERNEL

broke our mechanism for including initramfs kernel into live image.

As that image (hddimg) is used for flashing the Edison's MMC boot
partition, that in turn breaks the board boot as it's not able to
find the rootfs - it's on MMC and we build MMC/sdhc as modules right now.

Fixes #41.

Signed-off-by: Alex Tereschenko <alext.mkrs@gmail.com>