edison.env: add reboot sdhc and reboot usb
This allows unattended rebooting (i.e. without interrupting U-Boot manually)
from USB or SDHC. To make this work you need to have the rootfs with kernel
modules on the USB or SDHC and the kernel itself with built-in initramfs
in the FAT formated /dev/mmcblk0p9 disk (bzImage-initramfs for SDHC,
bzImage-initramfs-usb for USB).

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>