blink-led: when on acpi enabled kernel use libgpiod
On acpi enabled kernel pin numbering has changed. Make this script use libgpiod when acpi
is detected.

When no acpi is detected use the sysfs method. In this case remove changing the mode of
gpio40 through /sys/kernel/debug/gpio_debug as that is not supported any more.

On acpi this patch creates a rdepend on libgpiod-python bindings.
Also, python bindings require python3.

For this to work gpio40 needs to be in mode 0 and SPI_CLK_SEL and MUX18_DIR unused. Remove
spi in ACPI_FEATURES in edison.conf for this to happen.

While here fix some syntax changes introduced with python3.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>