initramfs: add modules for USB boot for kernel 4.20
With linux 4.20 usb is no longer forced to host. Instead pmic is used to read out the usb to determine
the mode. Add the required modules to initramfs to allow usb to powerup and allow switchroot to usb.
Of course when booting from usb gadget mode will not work at the same time.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>