post-install.sh: create /factory if non-existing
The /factory is mounted from a seperate partition
/dev/disk/by-partlabel/factory which is created during
manufacturing and contains a unique bluetooth_address
and serial_number. The partition is never removed during
flashing of images and not even when using flashall --recover.

In case for some reason the partion is lost, this will create
dummy files for both. The user can later manually edit the
files as needed.

The serial_number is on the label on the Intel Edison, but
the bluetooth_address needs to be reclaimed from backup.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>