btrfs: btrfsSnapshot.py: postprocess image and take snapshot
Yocto generates btrfs rootfs with everything in the top volume.
However, we want a subvolume @.

Further we want to take a read only snapshot @new and send that
to a file. Later we can send the snapshot file directly to edison
using ssh to implement an alternative OTA update.

Unfortunately it appears the needed image manipulations require
in part 1) root permissions and can not be added to the recipe.

This python script performs the needed manipulations at postbuild
time. It requires root permissions while taking and sending
snapshots and while removing dirs from the root. Some precautions
have been implemented to prevent accidentally deleting the root
of the host.

In addition we fixup fstab for booting btrfs and compress the snap to
prevent wasting space and enable sending to edison faster.

1) Taking snapshots require user ownership of the root directory
which can be implemented by creating a ext partition and then
converting to btrfs. However `btrfs send` always requires root
permissions.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>