btrfs: fstab: create btrfs compatible fstab
This creates a btrfs compatible fstab.btrfs. This one should be used after
ext4 conversion to btrfs and should also replace fstab after generating
the btrfs image.

Although it seems to make sense to call the btrfs volume rootfs, it is not. Root
is found on the kernel command line and mounted by the initramfs. But when we
boot an ext4 image from sdhc it looks for partlabel home and needs to find it
to mount /home. Else it will fail. So we will call the btrfs partlabel home.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>