edison-env: add rescue boot
We have initrd (built by recipe core-image-minimal-initramfs.bbappend) which provides
tools to repair and modify file systems. Normally you will enter a shell in the initrd
when the rootfs is not found within `debugshell` (default 30) seconds.

This add the command `run do_rescue` that will enter the initrd shell immediately.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>