btrfs.env: enable booting alternate kernel
The btrfs image allows to install a second kernel. This patch allows to tell U-Boot to
boot the alternate kernel.

The primary kernel and initrd are called 'bzImage' and 'initrd'. The alternate kernel and initrd
are called 'bzImageNew' and 'initrdNew'.

Setting `fw_setenv switch 1` from linux will cause U-Boot to add 'New' to the kernel and initrd
name and set 'boot_count' to 1. If the subsequent boot succeeds 'boot_count' will be set to 1 in
linux and allow future boots to continue with the New kernel. However if boot fail the kernel will
be reset to the default one again.

Therefore to force switching to the new kernel, in linux:
`fw_setenv switch 1 && reboot`.

To force switching to the default kernel, in linux:
`fw_setenv boot_count 1 && reboot`.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>