linux-yocto:revert "linux-yocto: replace TX lineair buffer by scatter-gather"
This reverts bffb7ef4edfb55adade3aa233b546d2de64a0480

We found that that the patch "0001-WIP-serial-8250_dma-use-sgl-on-transmit.patch"
works for many work loads. But in case of pppd which push the transmit buffer full
continously after a short while transmit hangs. When the hang occurs it seems
that dma fills TX FIFO the second time and in total 128 bytes are sent. But for some
reason no more DMA transfers to the FIFO are initiated. Since no more transfers are
triggered TX stalls, but also DMA descriptor never terminates. Only reboot resolves
the hang.

The mystery is even greater since the linear buffer transfer behind the scenes uses
a sg list of length 1.

Until we find out what cause this revert the sg code.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>