linux-yocto: disable DMA_API_DEBUG by default
Up to now we had DMA_API_DEBUG enabled for driver testing. However, this appears
to come with a performance hit. Moreover, it appears that it raises false WARN
in some cases, in particular in the usb hub driver. Here a buffer and a status
are kmalloc() in our case on the same cache line. But as x86 is cache coherent
this is not an issue on x86. Other platforms should do proper alignment or bounce
buffer to prevent the WARN. The false WARN on x86 may be taken care of upstream
but does not cause any harm.

Enable DMA_API_DEBUG if you are doing driver development and using DMA.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>