bluetooth_bd_addr: fix bluez tools hang when stdin is /dev/null
Bluetooth tools mtmgmt and bdaddr hang when run from an environment where
stdin is /dev/null as is the case by default in recent versions of systemd.
It is unclear if this is a bug (in Bluetoothtools) or intentional.
In any case resolve by letting systemd attach a tty to stdin as was the
previous behavior.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>