INIT_MANAGER: get rid of SysV traces
It appears somewhere since Dunfell selection of init manager was
simplified, but we missed that. As result SysV initscripts
are also being installed. Now it seems these scripts are conflicting
systemd in some cases (umountfs and umountnfs).
Setting INIT_MANAGER to systemd fixes this and also sets certain
DISTRO_FEATURES that we do no longer need to explicitely set.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>