fstab: do not mount home partition automatically
In the past having subvol=/@home as mount option did not stop the mount
if the partition happend to be ext4. So, the fstab line covered most use
cases.

Currently if home does not have a file system of the file system is broken
or the files system is ext4 systemd shutsdown and we run into a boot loop.

Although this can be fixed manually by doing `run do_rescue` from U-Boot it
makes more sense not to mount boot automatically and leave it to the user to
edit fstab as he sees fit.

With the original Intel Edison image this did not work as the home dir for root
was on /home/root. But since Zeus we already changed that to /root.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>